backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart
python-dotenv
pydantic>=2.5.0
pydantic-settings>=2.1.0
Pillow
google-generativeai
aiofiles
httpx
redis
celery
openpyxl
pandas
python-jose[cryptography]
passlib[bcrypt]
websockets
opencv-python-headless
numpy
pytesseract
easyocr
